+++ iroh-gateway/src/lib.rs
@@ -3,5 +3,6 @@ pub mod config;
 mod constants;
 pub mod core;
 mod error;
+mod headers;
 pub mod metrics;
 mod response;
